level refactoring
