+++ game2020/game2020/Interfaces/ITiles.cs
@@ -1,0 +1,12 @@
+﻿using game2020.World;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace game2020.Interfaces
+{
+    public interface ITiles
+    {
+        public CollisionTiles CollisionTiles { get; set; }
+    }
+}
